docker: use simple `go build` instead of bash script
The script assumes that it's running in a Git repository, but `docker
build` doesn't pass the Git history into the context, so it confuses the
script.

Unfortunately, that means running `caddy -version` with the resulting
binary won't return any useful information, but I don't think there is
any way to get the tag name or commit reference from the automated build
process on hub.docker.com.